# Improvements and optimizations to Aessafe