A few small cleanups and some optimizations to Aessafe.

Before:
```
test aes::bench::aes_safe_bench                   ... bench:       899 ns/iter (+/- 118) = 17 MB/s
test aes::bench::aes_safe_x8_bench                ... bench:      2146 ns/iter (+/- 219) = 59 MB/s
```

After:
```
test aes::bench::aes_safe_bench                   ... bench:       836 ns/iter (+/- 17) = 19 MB/s
test aes::bench::aes_safe_x8_bench                ... bench:      1016 ns/iter (+/- 77) = 125 MB/s
```